1.1.5 refactoring
- separation routes.py to routes
- fix button overlay on sentence (hope)
