+++ app/static/css/content.css
@@ -95,7 +95,7 @@ body {
     position: absolute;
     bottom: 20vh;
     left: 50%;
-    transform: translate(-50%, -50%);
+    transform: translate(-50%, 50%);
     display: flex;
     z-index: 10;
     width: 100%;
